fix: incorrect path causes subsequent crash during file removal
A regression caused by Commit afa17db changes the variable "path" unintended. This leads to the path pointing to the directory, and not the file. Subsequetly, the app crashes when trying to delete the directory, instead of the file.

The "path" here should be a second independent path.

Minified error, which is fixed by this commit:
line 150, in delete_file
    os.remove(file)
    ~~~~~~~~~^^^^^^
IsADirectoryError: [Errno 21] Is a directory